+++ source/CHANGELOG.md
@@ -11,7 +11,7 @@ Andreas Abel <andreas.abel@gu.se>  (unreleased)
 ## Bug fixes and small improvements
 
 * Haskell-GADT: generated code is warning free [#346]
-* Haskell: fixes in layout preprocessor [#343,#344,#345,#352,#353,#356]
+* Haskell: fixes in layout preprocessor [#343,#344,#345,#352,#353]
 * Haskell: print `[Char]` correctly, removed method `prtList` [#359]
 * Haskell: added missing import [#368], regression introduced in 2.9.1 by [#331]
 * C: fixed a space leak when parsing from a string in memory [#347]